Add mean distance and refactor PID interfaces.
Replaced the ManualPIDOut class with a single header that defines two classes: ManualPIDOutput, and TwoEncoders.
TwoEncoders allows for two encoders to be averaged together to get the actual distance the robot has travelled.
This will improve accuracy of distance PID, as well as allow new operations such as turning while ensuring the robot stays in place.